add codeowners file
also moved calc constants into lib/ and lerp() into lib/Math.ts to fit a bit more nicely
