+++ src/lib/Math.ts
@@ -1,1 +1,10 @@
 export type Factor = [factor: number, divisor: number];
+
+export const lerp = (curr: number, srcStart: number, srcEnd: number, dstStart: number, dstEnd: number): number => {
+  // does this need to be int-based?
+  const srcRange = srcEnd - srcStart;
+  const dstRange = dstEnd - dstStart;
+  const currNorm = curr - srcStart;
+
+  return Math.trunc(((currNorm * dstRange) / srcRange) + dstStart);
+};
